fix: Fix spotless problem
