+++ common/src/main/java/com/b1nd/dodam/common/network/di/ApplicationScope.kt
@@ -22,6 +22,5 @@ object CoroutineScopesModule {
     @Provides
     @Singleton
     @ApplicationScope
-    fun providesCoroutineScope(@Dispatcher(DispatcherType.Default) dispatcher: CoroutineDispatcher): CoroutineScope =
-        CoroutineScope(SupervisorJob() + dispatcher)
+    fun providesCoroutineScope(@Dispatcher(DispatcherType.Default) dispatcher: CoroutineDispatcher): CoroutineScope = CoroutineScope(SupervisorJob() + dispatcher)
 }
