--- data/wakeup-song/src/main/java/com/b1nd/dodam/wakeup_song/di/RepositoryModule.kt
@@ -1,18 +1,0 @@
-package com.b1nd.dodam.wakeup_song.di
-
-import com.b1nd.dodam.wakeup_song.WakeupSongRepository
-import com.b1nd.dodam.wakeup_song.repository.WakeupSongRepositoryImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface RepositoryModule {
-
-    @Binds
-    fun bindsWakeupSongRepository(
-        wakeupSongRepositoryImpl: WakeupSongRepositoryImpl
-    ): WakeupSongRepository
-}
