+++ data/wakeup-song/src/main/java/com/b1nd/dodam/wakeupsong/repository/WakeupSongRepositoryImpl.kt
@@ -1,0 +1,34 @@
+package com.b1nd.dodam.wakeupsong.repository
+
+import com.b1nd.dodam.common.Dispatcher
+import com.b1nd.dodam.common.DispatcherType
+import com.b1nd.dodam.common.result.Result
+import com.b1nd.dodam.common.result.asResult
+import com.b1nd.dodam.model.WakeupSong
+import com.b1nd.dodam.wakeupsong.WakeupSongRepository
+import com.b1nd.dodam.wakeupsong.datasource.WakeupSongDataSource
+import com.b1nd.dodam.wakeupsong.mapper.toModel
+import javax.inject.Inject
+import kotlinx.collections.immutable.ImmutableList
+import kotlinx.collections.immutable.toImmutableList
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+
+internal class WakeupSongRepositoryImpl @Inject constructor(
+    private val network: WakeupSongDataSource,
+    @Dispatcher(DispatcherType.IO) private val dispatcher: CoroutineDispatcher,
+) : WakeupSongRepository {
+    override fun getAllowedWakeupSongs(year: Int, month: Int, day: Int): Flow<Result<ImmutableList<WakeupSong>>> {
+        return flow {
+            emit(
+                network.getAllowedWakeupSongs(year, month, day)
+                    .map { it.toModel() }
+                    .toImmutableList(),
+            )
+        }
+            .asResult()
+            .flowOn(dispatcher)
+    }
+}
