+++ designsystem/src/main/java/com/b1nd/dodam/designsystem/theme/Iconography.kt
@@ -16,8 +16,8 @@ import com.b1nd.dodam.designsystem.icons.Cancel
 import com.b1nd.dodam.designsystem.icons.Error
 import com.b1nd.dodam.designsystem.icons.Eye
 import com.b1nd.dodam.designsystem.icons.EyeSlash
-import com.b1nd.dodam.designsystem.icons.Home
 import com.b1nd.dodam.designsystem.icons.ForkAndKnife
+import com.b1nd.dodam.designsystem.icons.Home
 import com.b1nd.dodam.designsystem.icons.More
 import com.b1nd.dodam.designsystem.icons.Out
