+++ designsystem/src/main/java/com/b1nd/dodam/designsystem/theme/Theme.kt
@@ -79,6 +79,7 @@ fun DodamTheme(
             }
 
             darkTheme -> DarkColorScheme
+
             else -> LightColorScheme
         }
     val view = LocalView.current
