+++ dodam-student/build.gradle.kts
@@ -40,4 +40,5 @@ dependencies {
     implementation(projects.featureStudent.main)
     implementation(projects.feature.register)
     implementation(projects.feature.login)
+    implementation(projects.datastore)
 }
