--- dodam-student/src/main/kotlin/com/b1nd/dodam/student/DodamApplication.kt
@@ -1,7 +1,0 @@
-package com.b1nd.dodam.student
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class DodamApplication : Application()
