+++ dodam-student/src/main/kotlin/com/b1nd/dodam/student/util/DodamApplication.kt
@@ -1,0 +1,7 @@
+package com.b1nd.dodam.student.util
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class DodamApplication : Application()
