+++ feature-student/home/src/main/java/com/b1nd/dodam/student/home/HomeViewModel.kt
@@ -6,19 +6,18 @@ import androidx.lifecycle.viewModelScope
 import com.b1nd.dodam.common.result.Result
 import com.b1nd.dodam.data.meal.MealRepository
 import com.b1nd.dodam.student.home.model.HomeUiState
-import com.b1nd.dodam.wakeup_song.WakeupSongRepository
+import com.b1nd.dodam.wakeupsong.WakeupSongRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import javax.inject.Inject
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
     private val mealRepository: MealRepository,
-    private val wakeupSongRepository: WakeupSongRepository
+    private val wakeupSongRepository: WakeupSongRepository,
 ) : ViewModel() {
     private val _uiState = MutableStateFlow(HomeUiState())
     val uiState = _uiState.asStateFlow()
@@ -35,14 +34,16 @@ class HomeViewModel @Inject constructor(
                                     meal = Triple(
                                         result.data.breakfast ?: "오늘은 아침이 없어요.",
                                         result.data.lunch ?: "오늘은 점심이 없어요.",
-                                        result.data.dinner ?: "오늘은 저녁이 없어요."
-                                    )
+                                        result.data.dinner ?: "오늘은 저녁이 없어요.",
+                                    ),
                                 )
                             }
+
                             is Result.Error -> Log.e("getMeal", result.exception.message.toString())
+
                             is Result.Loading -> _uiState.update {
                                 it.copy(
-                                    isLoading = true
+                                    isLoading = true,
                                 )
                             }
                         }
@@ -55,13 +56,15 @@ class HomeViewModel @Inject constructor(
                             is Result.Success -> _uiState.update {
                                 it.copy(
                                     isLoading = false,
-                                    wakeupSongs = result.data
+                                    wakeupSongs = result.data,
                                 )
                             }
+
                             is Result.Error -> Log.e("getWakeupSong", result.exception.message.toString())
+
                             is Result.Loading -> _uiState.update {
                                 it.copy(
-                                    isLoading = true
+                                    isLoading = true,
                                 )
                             }
                         }
