+++ feature-student/main/src/main/java/com/b1nd/dodam/student/main/navigation/MainNavigation.kt
@@ -2,7 +2,6 @@ package com.b1nd.dodam.student.main.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavHostController
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
