+++ network/wakeup-song/build.gradle.kts
@@ -24,4 +24,5 @@ dependencies {
     testImplementation(libs.ktor.client.mock)
     testImplementation(libs.junit)
     testImplementation(libs.kotlinx.coroutines.test)
+    androidTestImplementation(libs.junit)
 }
