--- network/wakeup-song/src/main/java/com/b1nd/dodam/wakeup_song/di/DataSourceModule.kt
@@ -1,17 +1,0 @@
-package com.b1nd.dodam.wakeup_song.di
-
-import com.b1nd.dodam.wakeup_song.api.WakeupSongService
-import com.b1nd.dodam.wakeup_song.datasource.WakeupSongDataSource
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface DataSourceModule {
-    @Binds
-    @Singleton
-    fun bindsWakeupSongDataSource(wakeupSongService: WakeupSongService): WakeupSongDataSource
-}
