--- network/wakeup-song/src/main/java/com/b1nd/dodam/wakeup_song/model/NetworkWakeupSongStatus.kt
@@ -1,10 +1,0 @@
-package com.b1nd.dodam.wakeup_song.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-enum class NetworkWakeupSongStatus {
-    ALLOWED,
-    PENDING,
-    DENIED
-}
