+++ network/wakeup-song/src/main/java/com/b1nd/dodam/wakeupsong/model/NetworkWakeupSongStatus.kt
@@ -1,0 +1,10 @@
+package com.b1nd.dodam.wakeupsong.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class NetworkWakeupSongStatus {
+    ALLOWED,
+    PENDING,
+    DENIED,
+}
